Recursively find directory contents
Fixes #9

This also makes it so this function only returns files. As far as I can
tell, this behavior change is OK since the extension filtering was
happening downstream anyway; directories were being dropped out there
already.

In the future, we could refactor this function and the filtering to
happen together. There's no reason this can't recursively find Haskell
files from the start. We've already made the [".", ".."] check redundant
in the filtering step, but I've left it there to keep the bug-fix diff
small.